Rework prevent duplicate clicks in web

+++ packages/web/src/components/SettingsPopupData.js
@@ -46,6 +46,12 @@ export const SettingsPopupData = _SettingsPopupData;
 
 class _SettingsPopupDataExport extends React.PureComponent {
 
+  constructor(props) {
+    super(props);
+
+    this.didClick = false;
+  }
+
   componentWillUnmount() {
     if (this.props.exportAllDataProgress) {
       const { total, done } = this.props.exportAllDataProgress;
@@ -54,7 +60,9 @@ class _SettingsPopupDataExport extends React.PureComponent {
   }
 
   onExportAllDataBtnClick = () => {
+    if (this.didClick) return;
     this.props.exportAllData();
+    this.didClick = true;
   }
 
   render() {
@@ -164,6 +172,8 @@ class _SettingsPopupDataDelete extends React.PureComponent {
       didCheckConfirm: false,
       isRequireConfirmShown: false,
     };
+
+    this.didClick = false;
   }
 
   componentWillUnmount() {
@@ -178,11 +188,14 @@ class _SettingsPopupDataDelete extends React.PureComponent {
   }
 
   onDeleteAllDataBtnClick = () => {
+    if (this.didClick) return;
+
     if (this.state.didCheckConfirm) {
       if (this.state.isRequireConfirmShown) {
         this.setState({ isRequireConfirmShown: false });
       }
       this.props.deleteAllData();
+      this.didClick = true;
       return;
     }
